fix(ollama): route readiness through Docker adapter
