+++ src/lib/adapters/docker/command.ts
@@ -1,0 +1,20 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import { run, runCapture } from "../../runner";
+
+export type DockerRunOptions = Parameters<typeof run>[1];
+export type DockerCaptureOptions = Parameters<typeof runCapture>[1];
+export type DockerRunResult = ReturnType<typeof run>;
+
+export function dockerArgv(args: readonly string[]): string[] {
+  return ["docker", ...args];
+}
+
+export function dockerRun(args: readonly string[], opts: DockerRunOptions = {}): DockerRunResult {
+  return run(dockerArgv(args), opts);
+}
+
+export function dockerCapture(args: readonly string[], opts: DockerCaptureOptions = {}): string {
+  return runCapture(dockerArgv(args), opts);
+}
